@@ -1,6 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
+        <TargetFramework>net9.0</TargetFramework>
     </PropertyGroup>
 
     <ItemGroup>
@@ -9,6 +9,6 @@
     </ItemGroup>
 
     <ItemGroup>
-        <PackageReference Include="Ertis.MongoDB.Queries" Version="7.0.9" />
+        <PackageReference Include="Ertis.MongoDB.Queries" Version="9.0.3" />
     </ItemGroup>
 </Project>
