@@ -1,25 +1,25 @@
 <Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
+        <TargetFramework>net9.0</TargetFramework>
         <PackageId>ErtisAuth.Core</PackageId>
-        <PackageVersion>7.7.5</PackageVersion>
+        <PackageVersion>9.0.5</PackageVersion>
         <Authors>Ertuğrul Özcan</Authors>
         <Description>ErtisAuth.Core</Description>
         <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
-        <Copyright>Ertuğrul Özcan | Copyright 2021 (c)</Copyright>
-        <PackageReleaseNotes>.Net 7</PackageReleaseNotes>
+        <Copyright>Ertuğrul Özcan | Copyright 2025 (c)</Copyright>
+        <PackageReleaseNotes>.Net 9</PackageReleaseNotes>
         <PackageTags>ErtisAuth.Core</PackageTags>
         <Title>ErtisAuth.Core</Title>
         <PackageProjectUrl>https://github.com/ertugrulozcan/ErtisAuth</PackageProjectUrl>
         <RepositoryUrl>https://github.com/ertugrulozcan/ErtisAuth</RepositoryUrl>
-        <AssemblyVersion>7.7.5</AssemblyVersion>
-        <FileVersion>7.7.5</FileVersion>
+        <AssemblyVersion>9.0.5</AssemblyVersion>
+        <FileVersion>9.0.5</FileVersion>
     </PropertyGroup>
     
     <ItemGroup>
-        <PackageReference Include="Ertis.Core" Version="7.1.0" />
-        <PackageReference Include="Ertis.Schema" Version="7.3.8" />
-        <PackageReference Include="Ertis.Security" Version="7.0.5" />
+        <PackageReference Include="Ertis.Core" Version="9.0.1" />
+        <PackageReference Include="Ertis.Schema" Version="9.0.2" />
+        <PackageReference Include="Ertis.Security" Version="9.0.1" />
         <PackageReference Include="Newtonsoft.Json.Schema" Version="4.0.1" />
     </ItemGroup>
     
