@@ -1,23 +1,23 @@
 <Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
+        <TargetFramework>net9.0</TargetFramework>
         <PackageId>ErtisAuth.Extensions.Authorization</PackageId>
-        <PackageVersion>7.7.2</PackageVersion>
+        <PackageVersion>9.0.5</PackageVersion>
         <Authors>Ertuğrul Özcan</Authors>
         <Description>ErtisAuth.Extensions.Authorization</Description>
         <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
-        <Copyright>Ertuğrul Özcan | Copyright 2021 (c)</Copyright>
-        <PackageReleaseNotes>.Net 7</PackageReleaseNotes>
+        <Copyright>Ertuğrul Özcan | Copyright 2025 (c)</Copyright>
+        <PackageReleaseNotes>.Net 9</PackageReleaseNotes>
         <PackageTags>ErtisAuth.Extensions.Authorization</PackageTags>
         <Title>ErtisAuth.Extensions.Authorization</Title>
         <PackageProjectUrl>https://github.com/ertugrulozcan/ErtisAuth</PackageProjectUrl>
         <RepositoryUrl>https://github.com/ertugrulozcan/ErtisAuth</RepositoryUrl>
-        <AssemblyVersion>7.7.2</AssemblyVersion>
-        <FileVersion>7.7.2</FileVersion>
+        <AssemblyVersion>9.0.5</AssemblyVersion>
+        <FileVersion>9.0.5</FileVersion>
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Microsoft.AspNetCore.Authorization" Version="7.0.13" />
+        <PackageReference Include="Microsoft.AspNetCore.Authorization" Version="9.0.8" />
     </ItemGroup>
 
     <ItemGroup>
