@@ -1,19 +1,19 @@
 <Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
+        <TargetFramework>net9.0</TargetFramework>
         <PackageId>ErtisAuth.Extensions.Http</PackageId>
-        <PackageVersion>7.7.4</PackageVersion>
+        <PackageVersion>9.0.5</PackageVersion>
         <Authors>Ertuğrul Özcan</Authors>
         <Description>ErtisAuth.Extensions.Http</Description>
         <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
-        <Copyright>Ertuğrul Özcan | Copyright 2021 (c)</Copyright>
-        <PackageReleaseNotes>.Net 7</PackageReleaseNotes>
+        <Copyright>Ertuğrul Özcan | Copyright 2025 (c)</Copyright>
+        <PackageReleaseNotes>.Net 9</PackageReleaseNotes>
         <PackageTags>ErtisAuth.Extensions.Http</PackageTags>
         <Title>ErtisAuth.Extensions.Http</Title>
         <PackageProjectUrl>https://github.com/ertugrulozcan/ErtisAuth</PackageProjectUrl>
         <RepositoryUrl>https://github.com/ertugrulozcan/ErtisAuth</RepositoryUrl>
-        <AssemblyVersion>7.7.4</AssemblyVersion>
-        <FileVersion>7.7.4</FileVersion>
+        <AssemblyVersion>9.0.5</AssemblyVersion>
+        <FileVersion>9.0.5</FileVersion>
     </PropertyGroup>
 
     <ItemGroup>
@@ -25,6 +25,6 @@
     </ItemGroup>
 
     <ItemGroup>
-        <PackageReference Include="Ertis.TemplateEngine" Version="7.0.8" />
+        <PackageReference Include="Ertis.TemplateEngine" Version="9.0.1" />
     </ItemGroup>
 </Project>
